Configure build with rollup
This allows us to output different formats. At the moment it outputs
a UMD module and an ES module and it drops the AMD module. AMD modules
have been superseded in recent years by UMD and ES modules.

You can still use Idiomorph via a browser global by including the
UMD module, but you can also include it as an ES module and use
in a project with a JS bundler, like esbuild, webpack, etc.

The htmx extension is also included as a UMD module so you can require
it directly in a browser.