Fix issue with attribute morphing
This fixes an issue with attribute morphing when doing multiple morphing where some attributes were randomly being left over from the previous markup. Upon testing I found that looping through the `toAttributes` via a `of` loop was skipping over some attributes of the map entirely and that switching to looping through the array of values resolved this.

#18