fix attribute morphing removal
this is a variant on the great PR created by @DEfusion & @pbower here:

https://github.com/bigskysoftware/idiomorph/pull/20

I ported over the test demonstrating the issue.  I implemented the fix slightly differently however because copying the arrays of attributes proved to be much slower (the checkbox per test went from 2ms to 9ms, compared w/ 2ms for morphdom).

Instead, i flipped the loop that does the attribute removal to iterate from the back to the front, which eliminates the skipping problem on deletion in a `for..of` loop by putting the mutation "behind" the indexes being considered.

This has no discernable perf impact and fixes the test.

Fixes https://github.com/bigskysoftware/idiomorph/issues/18